Improved retry logic
lodash.throttle invokes the throttled function *both* on the leading and trailing edge of the specified time window by default which means we’d retry *twice* :/

This makes the throttled retry function to properly behave, by invoking the function once, after waiting the specified amount of time.

Also applied a similar fix for the reportProgress function.